int -> float

+++ C/submitter/submitter.py
@@ -126,7 +126,7 @@ def finalize(directory):
         for filename in glob(directory + '/*.out'):
             with open(filename, 'r') as input_f:
                 lines = input_f.read()
-                values.extend(map(lambda x: int(x), lines.split('\n')))
+                values.extend(map(lambda x: float(x), lines.split('\n')))
                 f.write(lines)
 
             os.remove(filename)
